chore: split OGC Classic from OGC API shared implementation seams (#1579)
