+++ src/Honua.Protocols.Ogc.Shared/Common/OgcFeatureIdentifierResolver.cs
@@ -11,7 +11,7 @@ using Honua.Core.Features.Query;
 using Honua.Core.Features.Shared.Models;
 using Honua.Core.Queries.Filters;
 
-namespace Honua.Protocols.Ogc.Api.Features.Services;
+namespace Honua.Protocols.Ogc.Common;
 
 internal static class OgcFeatureIdentifierResolver
 {
