+++ src/Honua.Protocols.OgcClassic/Honua.Protocols.OgcClassic.csproj
@@ -9,11 +9,11 @@
     <!--
       Classic OGC protocol assembly (WFS, WMS, WMTS, WCS). Carved out of
       Honua.Protocols.Ogc.Classic per the modularization Phase 1 plan.
-      Its Infrastructure dependencies live in Honua.Hosting; the shared OGC base lives
-      in Honua.Protocols.Ogc.Shared. The WFS 2.0 / WMS / WMTS capabilities reuse OGC API
-      Features adapter types, so this assembly references Honua.Protocols.OgcApi (the only
-      inter-protocol edge; both collapse into the single "Ogc" family for isolation). It
-      may not reference Honua.Server.
+      Its Infrastructure dependencies live in Honua.Hosting; the shared OGC base
+      (GeoJSON wire models, extent/geometry helpers) lives in Honua.Protocols.Ogc.Shared.
+      Classic and OgcApi are sibling protocol adapters over that shared seam: this
+      assembly must NOT reference Honua.Protocols.OgcApi (enforced by
+      OgcSiblingIsolationTests; see #1579) and may not reference Honua.Server.
     -->
     <NoWarn>$(NoWarn);CS1591;IL2104;IL2026;IL3002;IL3050;CA1041;CA1000;CA1862;IDE0051;IDE0052;IDE0005;CA1844;IL2075</NoWarn>
     <RootNamespace>Honua.Protocols.Ogc</RootNamespace>
@@ -26,7 +26,6 @@
     <ProjectReference Include="..\Honua.Hosting\Honua.Hosting.csproj" />
     <ProjectReference Include="..\Honua.ServiceDefaults\Honua.ServiceDefaults.csproj" />
     <ProjectReference Include="..\Honua.Protocols.Ogc.Shared\Honua.Protocols.Ogc.Shared.csproj" />
-    <ProjectReference Include="..\Honua.Protocols.OgcApi\Honua.Protocols.OgcApi.csproj" />
   </ItemGroup>
 
   <ItemGroup>
