+++ src/Honua.Protocols.OgcClassic/Wfs20/Services/Wfs20Handler.StoredQueries.cs
@@ -27,9 +27,6 @@ using Honua.Infrastructure.Models;
 using Honua.Infrastructure.Services;
 using Honua.Infrastructure.Validation;
 using Honua.Protocols.Ogc.Common;
-using Honua.Protocols.Ogc.Api.Features;
-using Honua.Protocols.Ogc.Api.Features.Models;
-using Honua.Protocols.Ogc.Api.Features.Services;
 using Honua.Protocols.Ogc.Classic.Wfs20.Models;
 using Honua.ServiceDefaults;
 using NetTopologySuite;
